// lib/api/generated/model/featureDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */
import type { FeatureType } from "./featureType";

export interface FeatureDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  key?: string | null;
  type?: FeatureType;
}