// lib/api/generated/model/createFeatureCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */
import type { FeatureType } from "./featureType";

export interface CreateFeatureCommand {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  key?: string | null;
  type?: FeatureType;
}