// lib/api/generated/model/featureType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export type FeatureType = (typeof FeatureType)[keyof typeof FeatureType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FeatureType = {
  Limit: "Limit",
  Boolean: "Boolean",
} as const;